+++ .scripts/json-documentation.ts
@@ -15,11 +15,12 @@ try {
         "packages/implementations/react/src/*/*/src/!(*.spec|*.stories).types.ts",
       ),
     ),
-    subComponents: glob.sync(
-      path.join(
-        "packages/implementations/react/src/*/*/src/components/*/!(*.spec|*.stories).types.ts",
-      ),
-    ),
+    // subComponents: glob.sync(
+    //   path.join(
+    //     "packages/implementations/react/src/*/*/src/components/*/!(*.spec|*.stories).types.ts",
+    //   ),
+    // ),
+    subComponents: [],
   };
 
   const options: DocgenOptions = {