+++ packages/implementations/react/src/atomic/Toast/src/context/Manager/manager.types.ts
@@ -1,0 +1,8 @@
+import { ToastProps } from "../../toast.types";
+
+export interface ContextProps {
+  closeToast: (id: string) => void;
+  addToast: (
+    toast: Omit<ToastProps, "autoClose" | "id"> & { id?: string },
+  ) => void;
+}