+++ packages/styles/src/packages/atomic/skeleton/rarui-skeleton.types.ts
@@ -1,3 +1,22 @@
-import { sprinkle } from "./rarui-skeleton.css";
+import {
+  Conditions,
+  AddDollar,
+  StandardLonghandProperties,
+} from "../../../index.types";
+import { borderRadiusProperties } from "../../../properties";
 
-export type SkeletonSprinkle = Parameters<typeof sprinkle>[0];
+type SkeletonDynamicProperties = Pick<
+  StandardLonghandProperties,
+  "width" | "height" | "maxWidth" | "maxHeight" | "minWidth" | "minHeight"
+>;
+
+export interface SkeletonSprinkle extends SkeletonDynamicProperties {
+  /**
+   * The borderRadius property sets the skeleton.
+   */
+  borderRadius?:
+    | AddDollar<keyof typeof borderRadiusProperties | string>
+    | Conditions<AddDollar<keyof typeof borderRadiusProperties>>
+    | string
+    | Conditions<string>;
+}