+++ packages/styles/src/packages/atomic/toast/rarui-toast.types.ts
@@ -1,0 +1,4 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { styles } from "./rarui-toast.css";
+
+export type ToastVariants = RecipeVariants<typeof styles.toast>;