+++ packages/styles/src/packages/atomic/toggle/index.ts
@@ -1,0 +1,6 @@
+import { styles } from "./rarui-toggle.css";
+
+export const toggle = {
+  classnames: { ...styles },
+};
+export type { ToggleVariants } from "./rarui-toggle.types";