+++ packages/styles/src/types/index.ts
@@ -1,0 +1,4 @@
+export type {
+  StandardLonghandProperties,
+  StandardShorthandProperties,
+} from "./css.types";