+++ packages/implementations/react/CHANGELOG.md
@@ -2,11 +2,28 @@
 
 `@rarui-react/components` components is a component library built with [React](https://reactjs.org).
 
-## 2024-04-18 `1.11.0`
+## 2024-04-18 `1.13.0`
 
 #### 🎉 New features
 
 - Added new atomic `Banner` component. ([#26](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/26) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
+
+## 2024-04-25 `1.12.0`
+
+#### 🎉 New features
+
+- Added new atomic `Toast` component. ([#25](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/25) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
+- Added new atomic `IconButton` component. ([#25](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/25) by [@junior](https://git.rarolabs.com.br/junior))
+
+## 2024-04-25 `1.11.0`
+
+#### 🐛 Bug fixes
+
+- Fixed typing of `Box`, `Button`, `Chip`, `Link`, `Text` and `Title` atomic components. ([#27](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/27) by [@junior](https://git.rarolabs.com.br/junior))
+
+#### 🎉 New features
+
+- Added new atomic `Toggle` component. ([#27](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/27) by [@junior](https://git.rarolabs.com.br/junior))
 
 ## 2024-04-18 `1.10.0`