+++ packages/implementations/react/src/atomic/Link/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Link component allows us to navigate to external addresses.
 
+## 2024-04-25 `1.0.1`
+
+#### 🐛 Bug fixes
+
+- Fixed component typing. ([#27](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/27) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-03-23 `1.0.0`
 
 ### 🎉 New features