+++ packages/styles/CHANGELOG.md
@@ -2,11 +2,28 @@
 
 RarUI Styles deprive all styles needed to build components.
 
-## 2024-04-25 `1.11.0`
+## 2024-04-26 `1.13.0`
 
 #### 🎉 New features
 
 - Added new style pack for `banner` component. ([#26](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/26) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
+
+## 2024-04-25 `1.12.0`
+
+#### 🎉 New features
+
+- Added new style pack for `toast` component. ([#25](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/25) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
+- Added new style pack for `iconButton` component. ([#25](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/25) by [@junior](https://git.rarolabs.com.br/junior))
+
+## 2024-04-25 `1.11.0`
+
+#### 🐛 Bug fixes
+
+- Adjusting the typing of all components that use rainbow sprinkle. ([#27](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/27) by [@junior](https://git.rarolabs.com.br/junior))
+
+#### 🎉 New features
+
+- Added new style pack for `toggle` component. ([#27](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/27) by [@junior](https://git.rarolabs.com.br/junior))
 
 ## 2024-04-18 `1.10.0`